EnemyUpgrade and more
Enemies can now be upgraded.
Created Hand of Fate to move and select upgrade; Still needs work.
Changed exp to a float variable.
Moved Upgrades enum to UpdateStats script.

+++ Project Survival/Assets/Scripts/ItemManager.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class ItemManager : MonoBehaviour
 {
     public Dictionary<string, int> skillLevelDict = new(); //Saves skill level here to remember.
-    public Dictionary<string, int> skillExpDict = new(); //Saves the exp of skills here.
+    public Dictionary<string, float> skillExpDict = new(); //Saves the exp of skills here.
     public Dictionary<DraggableItem, int> skillOrbList = new(); //skill orb inventory
     public Dictionary<DraggableItem, int> skillGemList = new(); //skill gem inventory. Holds the gem in inventory and the total amount.
     public List<ItemDescription> availablePItemList = new(); //Passive Item shop inventory. Item is removed from list when there is no more quantity left.
